fix(demo.R): Workaround the readLines() error.
The `demo.R` test script that ships with blueogive/mro-docker includes
a line like this:
```r
source(url("http://some/url/filename.R"))
```
which throws an error when run within the RStudio console of this
Docker image. MRO runs the same line without complaint; it is only when
RStudio is part of the stack that the problem arises. I can trace the
issue to `.Interal(readLines())` but no further.

Replacing the line with this works around the problem:
```r
source(curl::curl("http://some/url/filename.R"))
```